requirements.txt
ccxt
pandas
numpy
matplotlib
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
httptools
websockets
pyinstaller
yfinance
requests
feedparser
pytrends
lxml
